[Cleanup] Rename .react.js components as .js.jsx for consistency
