+++ src/main/resources/assets/javascripts/components/ConnectionErrors.js.jsx
