+++ src/main/resources/assets/javascripts/components/DataPreview.js.jsx
