+++ src/main/resources/assets/javascripts/components/MetaDataPreview.js.jsx
