+++ src/main/resources/assets/javascripts/components/SearchInputField.js.jsx
